Added markers to my OxyPlot example. Got rid of the _Click Event handler in the code-behind MainWindow by replacing it with a class that implemented the ICommand interface. Note that whereas I was showing a MessageBox before, I am no longer launching anyting on the UI directly in my code that implements ICommand.
This article - https://www.codeproject.com/Articles/1052346/ICommand-Interface-in-WPF - did a good job of showing a.) how to create the code that implmeneted ICommand as well as how to create the code that implemented its partner, code that implements INotify PropertyChanged.

The anwser dated June 26, 2014 by Johan Larsson in this question - https://stackoverflow.com/questions/21914469/cant-update-oxyplot-graph-after-initialisation - was the key to getting the OxyPlot to update when points were added to Points_1.  Currently the InvalidateFlag is tied only to the number of points in Points_1.  When an OxyPlot contains multiple series, tying the InvalidateFlag to a sum of all points in all series seems reasonable (since each series can't have its own InvalidateFlag).